Added support for both NO and NC clutch switch on launch
In many car designs a NC clutch switch acts as NO switch because the pressure plate is forcing the switch open when clutch is released and vise versa for a NO clutch switch.  But in almost all designs I've seen they are always grounding.  A hot switching option may be added later for exceptions to that design standard.