Afroboltski's Fix for #655
Afroboltski's Fix for #655

There is a bug with the way Aux. input pins are initialised (and to be subsequently used for programmable I/O).

If an aux. input is assigned to pin x, the pinMode(pin, type) initialisation is done on pin x-1. This is not obvious because of the way the pin indices are defined in the INI file - DIGITAL_PIN is defined as strings starting from "1" as index 0, "2" as index 1, etc. hence why the offset is needed. But other defs use 0 as "Undefined" or "Disabled".

In the speeduino.ino file, we have

currentStatus.canin[currentStatus.current_caninchannel] = readAuxdigital((configPage9.Auxinpinb[currentStatus.current_caninchannel]&63)+1);

but this pin is initialised in sensors.ino,

byte pinNumber = (configPage9.Auxinpinb[currentStatus.current_caninchannel]&127); if( pinIsUsed(pinNumber) ) { BIT_SET(currentStatus.engineProtectStatus, PROTECT_IO_ERROR); } else { pinMode( pinNumber, INPUT); auxIsEnabled = true; }

causing a problem if a). pin 36 is already used (because we actually want pin 37, but the pinIsUsed check is done on pin 36), and b). this is the only aux. channel used, meaning that auxIsEnabled will never be set to true.

You may not even notice the problem if there were other aux inputs where x-1 didn't collide, and auxIsEnabled got set to true by "accident". Additionally the pinMode defaults to INPUT.

I believe byte pinNumber = (configPage9.Auxinpinb[currentStatus.current_caninchannel]&127); should be byte pinNumber = (configPage9.Auxinpinb[currentStatus.current_caninchannel]&63) + 1;