Speed up CLT by miniature amount
Set CLT, fuelpressure and oilpressure reading intervals to 240ms. Prevent this from being too close to the maximum 255, so that it do not skip readings when loop speed happens to be extremely low. Now it do not skip readings when main loop takes up to  15ms during time is up period(read should happen). As low as 66loops/s but must take account some loops are quicker than others.
Logic analyser tested on atmega2560 for not having misses.