Tweak tests.
Add check to prevent testing channelDegrees on channels that are inactive and so not are reguired to be initialised.
Also increase startup delay for decoders test because it seems to be more stable this way, at least with my particular board. With less delay some first tests tend to go missing or be replaced with something else.